Move DEVICE_WARNINGS constant
