+++ src/tuya_device_handlers/device_wrapper/__init__.py
@@ -1,11 +1,9 @@
 """Tuya device wrapper."""
 
 from .base import DeviceWrapper
-from .const import DEVICE_WARNINGS
 from .exception import SetValueOutOfRangeError
 
 __all__ = [
-    "DEVICE_WARNINGS",
     "DeviceWrapper",
     "SetValueOutOfRangeError",
 ]
