--- src/tuya_device_handlers/device_wrapper/const.py
@@ -1,5 +1,0 @@
-"""Tuya device wrapper."""
-
-# Dictionary to track logged warnings to avoid spamming logs
-# Keyed by device ID
-DEVICE_WARNINGS: dict[str, set[str]] = {}
